// src/app/models/oferta.ts
export class oferta{
    Error: string;
    Key: string;
    CargoBasicoActual: number;
    CargoBasicoOferta: number;
    CloudActual: string;
    CloudOferta: string;
    CodAbonado: string;
    CodigoPlanOferta: string;
    CodigoPlanVozActual: string;
    Corte: number;
    DataActual:string;
    DataOferta: string;
    DatosCompartidoOferta: string;
    DatosCompartidosActual: string;
    DescripcionPlanActual: string;
    DescripcionPlanOferta: string;
    Documento: string;
    DocumentoTipo:string;
    EstadoCliente: string;
    FacebookActual: string;
    FacebookOferta: string;
    FamiliayAmigosActual: string;
    FamiliayAmigosOferta: string;
    FechaActivacion: string;
    LatamActual:string;
    LatamOferta: string;
    LineActual: string;
    LineOferta: string;
    MinutosDesdeActual: string;
    MinutosDesdeOferta: string;
    MinutosMovistarActual: string;
    MinutosMovistarOferta:string;
    MinutosOtrosActual: string;
    MinutosOtrosOferta: string;
    MovistarPlayActual: string;
    MovistarPlayOferta:string;
    NapsterActual: string;
    NapsterOferta: null;
    SmSActual:string;
    SmsOferta: string;
    TelefonoContratado: string;
    TwitterActual: string;
    TwitterOferta: string;
    WazeActual: string;
    WazeOferta: string;
    WhatsappActual: string;
    WhatsappOferta: string;
}